+++ client/index.html
@@ -18,6 +18,6 @@
 	</head>
 	<body>
 		<div id="root"></div>
-		<script type="module" src="/src/Main.jsx"></script>
+		<script type="module" src="/src/Index.jsx"></script>
 	</body>
 </html>
