+++ client/src/components/screens/NotFound.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export const NotFound = () => {
+  return (
+    <>
+      <h3>Sorry the page you are looking for isn't available.</h3>
+    </>
+  )
+}
