+++ client/src/index.css
@@ -1,17 +1,24 @@
+.innerContainer {
+	display: flex;
+	position: absolute;
+	width: 100%;
+	top: 72px;
+	bottom: 64px;
+}
+
 .map-controls {
 	position: absolute;
 	display: flex;
 	bottom: 0;
 	left: 0;
 	align-items: center;
-	justify-content: center;
-	width: 90vw;
+	justify-content: space-between;
+	width: 100vw;
 }
 
 .map {
 	width: 100vw;
-	height: 100vh;
-	top: 60px;
+	height: 85vh;
 	cursor: grab;
 }
 
@@ -19,7 +26,6 @@
 	position: absolute;
 	display: flex;
 	flex-direction: column;
-	top: 60px;
 	right: 0;
 }
